refactor: replace hooks in services
